\docType{package}
\name{MODEST}
\alias{MODEST}
\alias{MODEST-package}
\title{MODEST}
\description{
MODEST integrates results across replicated STRUCTURE runs
spanning a large rang of K values and therefore takes into
account multimodality of data. change
}
\details{
More info to come
}
\author{
Petri Kemppainen, \email{petrikemppainen2@gmail.com} and
Stuart Baird, \email{stuartj.e.baird@gmail.com}
}

