\name{plotKDAG}
\alias{plotKDAG}
\title{Plots KDAG}
\usage{
plotKDAG(KDAG, ...)
}
\arguments{
  \item{KDAG}{0utput \code{KDAG} from
  \code{\link{Qcols2KDAG}}}

  \item{...}{Additional graphical parameters to be passed
  on the the plot.igraph functions}
}
\description{
Plots output from \code{\link{Qcols2KDAG}}
}
\details{
More info to come
}
\author{
Petri Kemppainen, \email{petrikemppainen2@gmail.com} and
Stuart Baird, \email{stuartj.e.baird@gmail.com}
}
\keyword{plotDAG}

